requirements.txt
PyQt6>=6.4.0
argon2-cffi>=21.3.0
pycryptodome>=3.18.0